requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.7
Django==5.0
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docutils==0.20.1
drf-yasg==1.21.7
inflection==0.5.1
iniconfig==2.0.0
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.4
pytest-django==4.7.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
setuptools==69.0.3
sqlparse==0.4.4
ua-parser==0.18.0
uritemplate==4.1.1
user-agents==2.2.0
